requirements.txt
fastapi
uvicorn
gunicorn
pydantic
requests
scikit-learn
networkx
chromadb
sentence-transformers
faiss-cpu
moviepy
pillow
gTTS
scrapy
scrapy-playwright
